requirements.txt
streamlit
sympy
numpy
Pillow